+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `DUO137`: lint for insecure itsdangerous kwarg usage ([#36](https://github.com/duo-labs/dlint/issues/36))
 
+### Fixed
+- False positive for `DUO137`. ([#37](https://github.com/duo-labs/dlint/issues/39))
+
 ## [0.9.2] - 2019-11-21
 ### Fixed
 - False negative with arbitrary depth from import alias in bad module attribute ([#32](https://github.com/duo-labs/dlint/issues/32))