+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `DUO137`: lint for insecure itsdangerous kwarg usage ([#36](https://github.com/duo-labs/dlint/issues/36))
 
 ### Fixed
-- False positive for `DUO137`. ([#39](https://github.com/duo-labs/dlint/issues/39))
+- False positive for `DUO137` when kwarg missing ([#39](https://github.com/duo-labs/dlint/issues/39))
 
 ## [0.9.2] - 2019-11-21
 ### Fixed